Restructuring code as per review comments and internal discussions
